+++ alphastats/gui/utils/llm_helper.py
@@ -1,13 +1,16 @@
 import os
+import warnings
 from pathlib import Path
 from typing import List, Optional
 
 import pandas as pd
 import streamlit as st
+from dataset.plotting import plotly_object
+from plots.plot_utils import PlotlyObject
 
 from alphastats.gui.utils.analysis import NewAnalysisOptions
 from alphastats.gui.utils.state_keys import DefaultStates, LLMKeys, StateKeys
-from alphastats.llm.llm_integration import LLMIntegration, MessageKeys, Models
+from alphastats.llm.llm_integration import LLMIntegration, MessageKeys, Models, Roles
 from alphastats.llm.uniprot_utils import (
     ExtractedUniprotFields,
     format_uniprot_annotation,
@@ -364,3 +367,101 @@ def display_uniprot(
                     ],
                 )
             )
+
+
+@st.fragment
+def llm_chat(
+    llm_integration: LLMIntegration,
+    selected_analysis_key: str,
+    show_all: bool = False,
+    show_individual_tokens: bool = False,
+    show_prompt: bool = True,
+):
+    """The chat interface for the LLM analysis."""
+
+    # TODO dump to file -> static file name, plus button to do so
+    # Ideas: save chat as txt, without encoding objects, just put a replacement string.
+    # Offer bulk download of zip with all figures (via plotly download as svg.).
+    # Alternatively write it all in one pdf report using e.g. pdfrw and reportlab (I have code for that combo).
+
+    # TODO show errors by defualt
+    # e.g. {"result": "Error executing get_uniprot_info_for_search_string: 'st.session_state has no key "selected_uniprot_fields".
+
+    selected_analysis_session_state = st.session_state[StateKeys.LLM_CHATS][
+        selected_analysis_key
+    ]
+    model_name = llm_integration._model
+    # no. tokens spent
+    messages, total_tokens, pinned_tokens = llm_integration.get_print_view(
+        show_all=show_all
+    )
+    for message in messages:
+        with st.chat_message(message[MessageKeys.ROLE]):
+            st.markdown(message[MessageKeys.CONTENT])
+            if (
+                message[MessageKeys.PINNED]
+                or not message[MessageKeys.IN_CONTEXT]
+                or show_individual_tokens
+            ):
+                token_message = ""
+                if message[MessageKeys.PINNED]:
+                    token_message += ":pushpin: "
+                if not message[MessageKeys.IN_CONTEXT]:
+                    token_message += ":x: "
+                if show_individual_tokens:
+                    tokens = llm_integration.estimate_tokens(
+                        [message], model=model_name
+                    )
+                    token_message += f"*tokens: {str(tokens)}*"
+                st.markdown(token_message)
+            for artifact in message[MessageKeys.ARTIFACTS]:
+                if isinstance(artifact, pd.DataFrame):
+                    st.dataframe(artifact)
+                elif isinstance(
+                    artifact, (PlotlyObject, plotly_object)
+                ):  # TODO can there be non-plotly types here
+                    st.plotly_chart(artifact)
+                elif not isinstance(artifact, str):
+                    st.warning("Don't know how to display artifact:")
+                    st.write(artifact)
+
+    st.markdown(
+        f"*total tokens used: {str(total_tokens)}, tokens used for pinned messages: {str(pinned_tokens)}*"
+    )
+
+    if selected_analysis_session_state.get(LLMKeys.RECENT_CHAT_WARNINGS):
+        st.warning("Warnings during last chat completion:")
+        for warning in selected_analysis_session_state[LLMKeys.RECENT_CHAT_WARNINGS]:
+            st.warning(warning)  # str(warning.message).replace("\n", "\n\n"))
+
+    if show_prompt and (prompt := st.chat_input("Say something")):
+        with st.chat_message(Roles.USER):
+            st.markdown(prompt)
+            if show_individual_tokens:
+                st.markdown(
+                    f"*tokens: {str(llm_integration.estimate_tokens([{MessageKeys.CONTENT:prompt}], model=model_name))}*"
+                )
+        with st.spinner("Processing prompt..."), warnings.catch_warnings(
+            record=True
+        ) as caught_warnings:
+            llm_integration.chat_completion(prompt)
+            selected_analysis_session_state[LLMKeys.RECENT_CHAT_WARNINGS] = (
+                caught_warnings
+            )
+
+        selected_analysis_session_state[LLMKeys.RECENT_CHAT_WARNINGS].append(
+            "some warning"
+        )
+
+        st.rerun(scope="fragment")
+
+    st.download_button(
+        "Download chat log",
+        llm_integration.get_chat_log_txt(),
+        f"chat_log_{model_name}.txt",
+        "text/plain",
+    )
+
+    st.markdown(
+        "*icons: :pushpin: pinned message, :x: message no longer in context due to token limitations*"
+    )
